+++ integration/spark/spark3/src/main/java/io/openlineage/spark3/agent/lifecycle/plan/DataSourceV2RelationInputDatasetBuilder.java
@@ -7,6 +7,7 @@ package io.openlineage.spark3.agent.lifecycle.plan;
 
 import io.openlineage.client.OpenLineage;
 import io.openlineage.client.dataset.DatasetCompositeFacetsBuilder;
+import io.openlineage.spark.agent.util.DatasetVersionUtils;
 import io.openlineage.spark.api.AbstractQueryPlanInputDatasetBuilder;
 import io.openlineage.spark.api.DatasetFactory;
 import io.openlineage.spark.api.OpenLineageContext;
@@ -46,8 +47,8 @@ public class DataSourceV2RelationInputDatasetBuilder
   public List<OpenLineage.InputDataset> apply(DataSourceV2Relation relation) {
     DatasetCompositeFacetsBuilder datasetFacetsBuilder = factory.createCompositeFacetBuilder();
 
-    DatasetVersionDatasetFacetUtils.includeDatasetVersion(
-        context, datasetFacetsBuilder.getFacets(), relation);
+    DatasetVersionDatasetFacetUtils.extractVersionFromDataSourceV2Relation(context, relation)
+        .ifPresent(s -> DatasetVersionUtils.buildVersionFacets(context, datasetFacetsBuilder, s));
     return DataSourceV2RelationDatasetExtractor.extract(
         factory, context, relation, datasetFacetsBuilder);
   }