+++ integration/spark/spark3/src/test/java/io/openlineage/spark3/agent/lifecycle/plan/DataSourceV2ScanRelationOnEndInputDatasetBuilderTest.java
@@ -15,10 +15,10 @@ import static org.mockito.Mockito.when;
 import io.openlineage.client.OpenLineage;
 import io.openlineage.client.OpenLineage.InputDataset;
 import io.openlineage.spark.agent.lifecycle.SparkOpenLineageExtensionVisitorWrapper;
+import io.openlineage.spark.agent.util.DatasetVersionUtils;
 import io.openlineage.spark.api.DatasetFactory;
 import io.openlineage.spark.api.OpenLineageContext;
 import io.openlineage.spark3.agent.utils.DataSourceV2RelationDatasetExtractor;
-import io.openlineage.spark3.agent.utils.DatasetVersionDatasetFacetUtils;
 import java.util.List;
 import org.apache.spark.scheduler.SparkListenerApplicationEnd;
 import org.apache.spark.scheduler.SparkListenerApplicationStart;
@@ -83,8 +83,8 @@ class DataSourceV2ScanRelationOnEndInputDatasetBuilderTest {
 
     try (MockedStatic<DataSourceV2RelationDatasetExtractor> ignored =
         mockStatic(DataSourceV2RelationDatasetExtractor.class)) {
-      try (MockedStatic<DatasetVersionDatasetFacetUtils> facetUtilsMockedStatic =
-          mockStatic(DatasetVersionDatasetFacetUtils.class)) {
+      try (MockedStatic<DatasetVersionUtils> facetUtilsMockedStatic =
+          mockStatic(DatasetVersionUtils.class)) {
         when(DataSourceV2RelationDatasetExtractor.extract(
                 eq(factory), eq(context), eq(relation), any()))
             .thenReturn(datasets);
@@ -92,10 +92,7 @@ class DataSourceV2ScanRelationOnEndInputDatasetBuilderTest {
         Assertions.assertThat(builder.apply(scanRelation)).isEqualTo(datasets);
 
         facetUtilsMockedStatic.verify(
-            () ->
-                DatasetVersionDatasetFacetUtils.includeDatasetVersion(
-                    context, datasetFacetsBuilder, relation),
-            times(0));
+            () -> DatasetVersionUtils.buildVersionFacets(eq(context), any(), any()), times(0));
       }
     }
   }