+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/OpenLineage/OpenLineage/compare/1.25.0...HEAD)
 
+### Added
+* **Spark: Collect `ScanReport` and `CommitReport` in OpenLineage events when dealing with Iceberg tables.** [`#3256`](https://github.com/OpenLineage/OpenLineage/pull/3256) [@pawel-big-lebowski](https://github.com/pawel-big-lebowski)
+  *Collects additional Iceberg metrics for datasets read or written through the library. Visit `Dataset Metrics` docs for more details.*
+
 
 ## [1.25.0](https://github.com/OpenLineage/OpenLineage/compare/1.24.2...1.25.0) - 2024-11-25